Revert "added integer index (#34)"
This reverts commit cd7c01d95ceb60e5b10e3e2144afb32d7651af5a.
