+++ .github/workflows/go.yml
@@ -14,13 +14,18 @@ jobs:
     runs-on: ${{ matrix.os }}
     strategy:
       matrix:
-        go: [1.15]
+        go: [1.14]
         os: [ubuntu-latest, macos-latest, windows-latest]
     steps:
     - name: Setup Go
       uses: actions/setup-go@v1
       with:
         go-version: ${{ matrix.go }}
+    - name: Setup env
+      run: |
+        echo "::set-env name=GOPATH::$(go env GOPATH)"
+        echo "::add-path::$(go env GOPATH)/bin"
+      shell: bash
     - name: Set git to use LF
       # make sure that line endings are not converted on windows
       # as gofmt linter will report that they need to be changed
