+++ Makefile
@@ -1,5 +1,5 @@
 GO ?= go
-GOLANGCI_LINT ?= $$($(GO) env GOPATH)/bin/golangci-lint
+GOLANGCI_LINT ?= golangci-lint
 GOLANGCI_LINT_VERSION ?= v1.30.0
 GOGOPROTOBUF ?= protoc-gen-gogofaster
 GOGOPROTOBUF_VERSION ?= v1.3.1
@@ -26,7 +26,7 @@ lint: linter
 
 .PHONY: linter
 linter:
-	test -f $(GOLANGCI_LINT) || curl -sSfL https://raw.githubusercontent.com/golangci/golangci-lint/master/install.sh | sh -s -- -b $$($(GO) env GOPATH)/bin $(GOLANGCI_LINT_VERSION)
+	which $(GOLANGCI_LINT) || curl -sSfL https://raw.githubusercontent.com/golangci/golangci-lint/master/install.sh | sh -s -- -b $$($(GO) env GOPATH)/bin $(GOLANGCI_LINT_VERSION)
 
 .PHONY: vet
 vet:
