+++ cmd/dfs-cli/cmd/prompt.go
@@ -21,7 +21,6 @@ import (
 	"encoding/json"
 	"errors"
 	"fmt"
-
 	"github.com/fairdatasociety/fairOS-dfs/pkg/collection"
 
 	"io/ioutil"
