+++ pkg/collection/entry.go
@@ -17,10 +17,9 @@ limitations under the License.
 package collection
 
 type Manifest struct {
-	Name         string    `json:"name"`
-	IdxType      IndexType `json:"index_type"`
-	CreationTime int64     `json:"creation_time"`
-	Entries      []*Entry  `json:"entries,omitempty"`
+	Name         string   `json:"name"`
+	CreationTime int64    `json:"creation_time"`
+	Entries      []*Entry `json:"entries,omitempty"`
 	dirtyFlag    bool
 }
 
@@ -31,13 +30,10 @@ type Entry struct {
 	manifest *Manifest
 }
 
-func NewManifest(name string, time int64, idxType IndexType) *Manifest {
-	var entries []*Entry
+func NewManifest(name string, time int64) *Manifest {
 	return &Manifest{
 		Name:         name,
-		IdxType:      idxType,
 		CreationTime: time,
-		Entries:      entries,
 		dirtyFlag:    true,
 	}
 }
