+++ pkg/collection/index_test.go
@@ -17,180 +17,508 @@ limitations under the License.
 package collection_test
 
 import (
-	"encoding/json"
+	"bytes"
 	"errors"
 	"io/ioutil"
-	"strconv"
+	"net/http"
 	"testing"
 
 	"github.com/fairdatasociety/fairOS-dfs/pkg/account"
-	"github.com/fairdatasociety/fairOS-dfs/pkg/blockstore"
 	"github.com/fairdatasociety/fairOS-dfs/pkg/blockstore/bee/mock"
 	"github.com/fairdatasociety/fairOS-dfs/pkg/collection"
 	"github.com/fairdatasociety/fairOS-dfs/pkg/feed"
 	"github.com/fairdatasociety/fairOS-dfs/pkg/logging"
-	"github.com/fairdatasociety/fairOS-dfs/pkg/utils"
 )
 
 func TestIndex(t *testing.T) {
 	mockClient := mock.NewMockBeeClient()
 	logger := logging.New(ioutil.Discard, 0)
 	acc := account.New(logger)
-	ai := acc.GetUserAccountInfo()
 	_, _, err := acc.CreateUserAccount("password", "")
 	if err != nil {
 		t.Fatal(err)
 	}
 	fd := feed.New(acc.GetUserAccountInfo(), mockClient, logger)
-	user := acc.GetAddress(account.UserAccountIndex)
 
-	t.Run("create_index", func(t *testing.T) {
-		//  create an index
-		err := collection.CreateIndex("testdb0", "key", collection.StringIndex, fd, user, mockClient)
+	t.Run("close_and_open_index", func(t *testing.T) {
+		//  create and populate the index
+		err := collection.CreateIndex("testdb0", "key", fd, acc.GetAddress(account.UserAccountIndex), mockClient)
+		if err != nil {
+			t.Fatal(err)
+		}
+		index, err := collection.OpenIndex("testdb0", "key", fd, acc.GetUserAccountInfo(), acc.GetAddress(account.UserAccountIndex), mockClient, logger)
+		if err != nil {
+			t.Fatal(err)
+		}
+		kvMap := addLotOfDocs(t, index, mockClient)
+
+		// open the index again, simulating like another instance
+		index1, err := collection.OpenIndex("testdb0", "key", fd, acc.GetUserAccountInfo(), acc.GetAddress(account.UserAccountIndex), mockClient, logger)
 		if err != nil {
 			t.Fatal(err)
 		}
 
-		// check if the index is created
-		if !isIndexPresent(t, "testdb0", "key", fd, user, mockClient) {
-			t.Fatalf("index not found")
-		}
-	})
-
-	t.Run("create_already_present_index", func(t *testing.T) {
-		//  create an index
-		err := collection.CreateIndex("testdb1", "key", collection.StringIndex, fd, user, mockClient)
-		if err != nil {
-			t.Fatal(err)
+		for k, expectedValue := range kvMap {
+			gotValue := getDoc(t, k, index1, mockClient)
+			if !bytes.Equal(expectedValue, gotValue) {
+				t.Fatalf("expected expectedValue %s got expectedValue %s", expectedValue, gotValue)
+			}
 		}
 
-		//  create an index
-		err = collection.CreateIndex("testdb1", "key", collection.StringIndex, fd, user, mockClient)
-		if !errors.Is(err, collection.ErrIndexAlreadyPresent) {
-			t.Fatal(err)
-		}
-	})
-
-	t.Run("create_and_open_index", func(t *testing.T) {
-		//  create an index
-		err := collection.CreateIndex("testdb2", "key", collection.StringIndex, fd, user, mockClient)
-		if err != nil {
-			t.Fatal(err)
+		// check if anuuninserted value exists
+		gotValue := getDoc(t, "p", index1, mockClient)
+		if gotValue != nil {
+			t.Fatalf("found data for not inserted key")
 		}
 
-		//Open the index
-		_, err = collection.OpenIndex("testdb2", "key", fd, ai, user, mockClient, logger)
+		// delete the index
+		err = index1.DeleteIndex()
 		if err != nil {
 			t.Fatal(err)
 		}
 	})
 
-	t.Run("open_index_without_creating_it", func(t *testing.T) {
-		//Open the index
-		_, err = collection.OpenIndex("testdb3", "key", fd, ai, user, mockClient, logger)
-		if err != collection.ErrIndexNotPresent {
+	t.Run("get-doc", func(t *testing.T) {
+		//  create and populate the index
+		err := collection.CreateIndex("testdb1", "key", fd, acc.GetAddress(account.UserAccountIndex), mockClient)
+		if err != nil {
+			t.Fatal(err)
+		}
+		index, err := collection.OpenIndex("testdb1", "key", fd, acc.GetUserAccountInfo(), acc.GetAddress(account.UserAccountIndex), mockClient, logger)
+		if err != nil {
+			t.Fatal(err)
+		}
+		kvMap := addLotOfDocs(t, index, mockClient)
+
+		// get the expectedValue of keys and check against its actual expectedValue
+		for k, expectedValue := range kvMap {
+			gotValue := getDoc(t, k, index, mockClient)
+			if !bytes.Equal(expectedValue, gotValue) {
+				t.Fatalf("expected expectedValue %s got expectedValue %s", expectedValue, gotValue)
+			}
+		}
+
+		gotValue := getDoc(t, "p", index, mockClient)
+		if gotValue != nil {
+			t.Fatalf("found data for not inserted key")
+		}
+
+		// delete the index
+		err = index.DeleteIndex()
+		if err != nil {
 			t.Fatal(err)
 		}
 	})
 
-	t.Run("create_and_delete_index", func(t *testing.T) {
-		//  create an index
-		err := collection.CreateIndex("testdb4", "key", collection.StringIndex, fd, user, mockClient)
+	t.Run("get-count", func(t *testing.T) {
+		//  create and populate the index
+		err := collection.CreateIndex("testdb11", "key", fd, acc.GetAddress(account.UserAccountIndex), mockClient)
+		if err != nil {
+			t.Fatal(err)
+		}
+		index, err := collection.OpenIndex("testdb11", "key", fd, acc.GetUserAccountInfo(), acc.GetAddress(account.UserAccountIndex), mockClient, logger)
+		if err != nil {
+			t.Fatal(err)
+		}
+		kvMap := addLotOfDocs(t, index, mockClient)
+
+		// find the count
+		count, err := index.Count()
 		if err != nil {
 			t.Fatal(err)
 		}
 
-		//Open the index
-		idx, err := collection.OpenIndex("testdb4", "key", fd, ai, user, mockClient, logger)
-		if err != nil {
+		if uint64(len(kvMap)) != count {
 			t.Fatal(err)
 		}
 
-		// delete Index
-		err = idx.DeleteIndex()
+		// delete the index
+		err = index.DeleteIndex()
 		if err != nil {
 			t.Fatal(err)
 		}
 	})
 
-	t.Run("delete_index_without_creating_it", func(t *testing.T) {
-		// simulate index not present by creating and deleting it
-		err := collection.CreateIndex("testdb5", "key", collection.StringIndex, fd, user, mockClient)
+	t.Run("get-doc-del-doc-get-doc", func(t *testing.T) {
+		//  create and populate the index
+		err := collection.CreateIndex("testdb2", "key", fd, acc.GetAddress(account.UserAccountIndex), mockClient)
 		if err != nil {
 			t.Fatal(err)
 		}
-		idx, err := collection.OpenIndex("testdb5", "key", fd, ai, user, mockClient, logger)
+		index, err := collection.OpenIndex("testdb2", "key", fd, acc.GetUserAccountInfo(), acc.GetAddress(account.UserAccountIndex), mockClient, logger)
 		if err != nil {
 			t.Fatal(err)
 		}
-		err = idx.DeleteIndex()
-		if err != nil {
-			t.Fatal(err)
+		kvMap := addLotOfDocs(t, index, mockClient)
+
+		// get the value of the key just to check
+		value := kvMap["aa"]
+		gotValue := getDoc(t, "aa", index, mockClient)
+		if !bytes.Equal(value, gotValue) {
+			t.Fatalf("expected value %s got value %s", value, gotValue)
 		}
 
-		// delete Index which is not present
-		err = idx.DeleteIndex()
-		if err != collection.ErrIndexNotPresent {
-			t.Fatal(err)
+		// delete the key from the index
+		deletedValue := delDoc(t, "aa", index, mockClient)
+		if !bytes.Equal(value, deletedValue) {
+			t.Fatalf("expected value %s got value %s", value, gotValue)
 		}
-	})
 
-	t.Run("count_docs", func(t *testing.T) {
-		// create index and add some docs
-		err := collection.CreateIndex("testdb6", "key", collection.StringIndex, fd, user, mockClient)
+		// check if the key is not in the index
+		afterDeletedValue := getDoc(t, "aa", index, mockClient)
+		if afterDeletedValue != nil {
+			t.Fatalf("shuld not have got any value")
+		}
+
+		// delete the index
+		err = index.DeleteIndex()
 		if err != nil {
 			t.Fatal(err)
-		}
-
-		idx, err := collection.OpenIndex("testdb6", "key", fd, ai, user, mockClient, logger)
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		// add some documents
-		actualCount := uint64(100)
-		for i := 0; i < int(actualCount); i++ {
-			putDocInIndex(t, idx, "key"+strconv.Itoa(i), "value"+strconv.Itoa(i), collection.StringIndex)
-		}
-
-		// count and check the count
-		count, err := idx.CountIndex()
-		if err != nil {
-			t.Fatal(err)
-		}
-		if count != actualCount {
-			t.Fatalf("invalid count in index, expected %d got %d", actualCount, count)
 		}
 
 	})
 
+	t.Run("add-docs-iterrate", func(t *testing.T) {
+		//  create and populate the index
+		err := collection.CreateIndex("testdb3", "key", fd, acc.GetAddress(account.UserAccountIndex), mockClient)
+		if err != nil {
+			t.Fatal(err)
+		}
+		index, err := collection.OpenIndex("testdb3", "key", fd, acc.GetUserAccountInfo(), acc.GetAddress(account.UserAccountIndex), mockClient, logger)
+		if err != nil {
+			t.Fatal(err)
+		}
+		kvMap := addLotOfDocs(t, index, mockClient)
+
+		// create the iterator
+		itr, err := index.NewIterator("", "", 100)
+		if err != nil {
+			t.Fatal(err)
+		}
+
+		// iterate through the keys and check for the values returned
+		count := 0
+		for itr.Next() {
+			value := getDoc(t, itr.Key(), index, mockClient)
+			if !bytes.Equal(kvMap[itr.Key()], value) {
+				t.Fatalf("expected value %s but got %s for the key %s", string(kvMap[itr.Key()]), string(value), itr.Key())
+			}
+			count++
+		}
+
+		if len(kvMap) != count {
+			t.Fatalf("number of elements mismatch in iteration")
+		}
+
+		// delete the index
+		err = index.DeleteIndex()
+		if err != nil {
+			t.Fatal(err)
+		}
+	})
+
+	t.Run("add-docs-seek-iterate", func(t *testing.T) {
+		//  create and populate the index
+		err := collection.CreateIndex("testdb4", "key", fd, acc.GetAddress(account.UserAccountIndex), mockClient)
+		if err != nil {
+			t.Fatal(err)
+		}
+		index, err := collection.OpenIndex("testdb4", "key", fd, acc.GetUserAccountInfo(), acc.GetAddress(account.UserAccountIndex), mockClient, logger)
+		if err != nil {
+			t.Fatal(err)
+		}
+		kvMap := addLotOfDocs(t, index, mockClient)
+
+		// create the iterator
+		itr, err := index.NewIterator("abc", "bbb", 100)
+		if err != nil {
+			t.Fatal(err)
+		}
+
+		// iterate through the keys and check for the values returned
+		count := 0
+		for itr.Next() {
+			value := getValue(t, itr.Value(), mockClient)
+			if !bytes.Equal(kvMap[itr.Key()], value) {
+				t.Fatalf("expected value %s but got %s for the key %s", string(kvMap[itr.Key()]), string(value), itr.Key())
+			}
+			count++
+		}
+
+		if count != 7 {
+			t.Fatalf("number of elements mismatch in iteration")
+		}
+
+		// delete the index
+		err = index.DeleteIndex()
+		if err != nil {
+			t.Fatal(err)
+		}
+	})
+
+	t.Run("add-docs-seek-iterrate-with-limit", func(t *testing.T) {
+		//  create and populate the index
+		err := collection.CreateIndex("testdb5", "key", fd, acc.GetAddress(account.UserAccountIndex), mockClient)
+		if err != nil {
+			t.Fatal(err)
+		}
+		index, err := collection.OpenIndex("testdb5", "key", fd, acc.GetUserAccountInfo(), acc.GetAddress(account.UserAccountIndex), mockClient, logger)
+		if err != nil {
+			t.Fatal(err)
+		}
+		kvMap := addLotOfDocs(t, index, mockClient)
+
+		// create the iterator
+		itr, err := index.NewIterator("abc", "bbb", 5)
+		if err != nil {
+			t.Fatal(err)
+		}
+
+		// iterate through the keys and check for the values returned
+		count := 0
+		for itr.Next() {
+			value := getValue(t, itr.Value(), mockClient)
+			if !bytes.Equal(kvMap[itr.Key()], value) {
+				t.Fatalf("expected value %s but got %s for the key %s", string(kvMap[itr.Key()]), string(value), itr.Key())
+			}
+			count++
+		}
+
+		if count != 5 {
+			t.Fatalf("number of elements mismatch in iteration")
+		}
+
+		// delete the index
+		err = index.DeleteIndex()
+		if err != nil {
+			t.Fatal(err)
+		}
+	})
+
+	t.Run("batch-add-docs", func(t *testing.T) {
+		//  create and populate the index
+		err := collection.CreateIndex("testdb6", "key", fd, acc.GetAddress(account.UserAccountIndex), mockClient)
+		if err != nil {
+			t.Fatal(err)
+		}
+		index, err := collection.OpenIndex("testdb6", "key", fd, acc.GetUserAccountInfo(), acc.GetAddress(account.UserAccountIndex), mockClient, logger)
+		if err != nil {
+			t.Fatal(err)
+		}
+
+		// batch load and delete
+		batch, err := index.Batch()
+		if err != nil {
+			t.Fatal(err)
+		}
+
+		batchDocs := addBatchDocs(t, batch, mockClient)
+		err = batch.Write()
+		if err != nil {
+			t.Fatal(err)
+		}
+
+		// create the iterator
+		itr, err := index.NewIterator("", "", 100)
+		if err != nil {
+			t.Fatal(err)
+		}
+
+		// iterate through the keys and check for the values returned
+		count := 0
+		for itr.Next() {
+			value := getValue(t, itr.Value(), mockClient)
+			if !bytes.Equal(batchDocs[itr.Key()], value) {
+				t.Fatalf("expected value %s but got %s for the key %s", string(batchDocs[itr.Key()]), string(value), itr.Key())
+			}
+			count++
+		}
+
+		if len(batchDocs) != count {
+			t.Fatalf("number of elements mismatch in iteration")
+		}
+
+		// delete the index
+		err = index.DeleteIndex()
+		if err != nil {
+			t.Fatal(err)
+		}
+	})
+
+	t.Run("batch-add-del-docs", func(t *testing.T) {
+		//  create and populate the index
+		err := collection.CreateIndex("testdb7", "key", fd, acc.GetAddress(account.UserAccountIndex), mockClient)
+		if err != nil {
+			t.Fatal(err)
+		}
+		index, err := collection.OpenIndex("testdb7", "key", fd, acc.GetUserAccountInfo(), acc.GetAddress(account.UserAccountIndex), mockClient, logger)
+		if err != nil {
+			t.Fatal(err)
+		}
+
+		// batch load and delete
+		batch, err := index.Batch()
+		if err != nil {
+			t.Fatal(err)
+		}
+
+		batchDocs := addBatchDocs(t, batch, mockClient)
+		err = batch.Write()
+		if err != nil {
+			t.Fatal(err)
+		}
+
+		// create the iterator
+		itr, err := index.NewIterator("", "", 100)
+		if err != nil {
+			t.Fatal(err)
+		}
+
+		// iterate through the keys and check for the values returned
+		count := 0
+		for itr.Next() {
+			value := getValue(t, itr.Value(), mockClient)
+			if !bytes.Equal(batchDocs[itr.Key()], value) {
+				t.Fatalf("expected value %s but got %s for the key %s", string(batchDocs[itr.Key()]), string(value), itr.Key())
+			}
+			count++
+		}
+
+		if len(batchDocs) != count {
+			t.Fatalf("number of elements mismatch in iteration")
+		}
+
+		// delete the index
+		err = index.DeleteIndex()
+		if err != nil {
+			t.Fatal(err)
+		}
+	})
 }
 
-func isIndexPresent(t *testing.T, collectionName, indexName string, fd *feed.API, user utils.Address, client blockstore.Client) bool {
-	actualIndexName := collectionName + indexName
-	topic := utils.HashString(actualIndexName)
-	_, addr, err := fd.GetFeedData(topic, user)
-	if err == nil && len(addr) != 0 {
-		data, _, err := client.DownloadBlob(addr)
-		if err != nil {
-			return false
-		}
-		var manifest collection.Manifest
-		err = json.Unmarshal(data, &manifest)
-		if err != nil {
-			return false
-		}
-		if manifest.Name != actualIndexName {
-			return false
-		}
-		return true
-	}
-	return false
-}
-
-func putDocInIndex(t *testing.T, index *collection.Index, key, value string, idxTYpe collection.IndexType) {
-	err := index.Put(key, []byte(value), idxTYpe)
+func addDoc(t *testing.T, key string, value []byte, index *collection.Index, client *mock.MockBeeClient) {
+	ref, err := client.UploadBlob(value, false, false)
 	if err != nil {
-		t.Fatalf("could not add doc in index: %s:%s, %v", key, value, err)
+		t.Fatalf("could not add doc %s:%s, %v", key, value, err)
 	}
+	err = index.Put(key, ref)
+	if err != nil {
+		t.Fatalf("could not add doc in index: %s:%s, %v", key, ref, err)
+	}
+}
+
+func getDoc(t *testing.T, key string, index *collection.Index, client *mock.MockBeeClient) []byte {
+	ref, err := index.Get(key)
+	if err != nil {
+		if errors.Is(err, collection.ErrEntryNotFound) {
+			return nil
+		}
+		t.Fatal(err)
+	}
+	data, respCode, err := client.DownloadBlob(ref)
+	if err != nil {
+		t.Fatal(err)
+	}
+	if respCode != http.StatusOK {
+		t.Fatalf("invalid response code")
+	}
+	return data
+}
+
+func getValue(t *testing.T, ref []byte, client *mock.MockBeeClient) []byte {
+	data, respCode, err := client.DownloadBlob(ref)
+	if err != nil {
+		t.Fatal(err)
+	}
+	if respCode != http.StatusOK {
+		t.Fatalf("invalid response code")
+	}
+	return data
+}
+
+func delDoc(t *testing.T, key string, index *collection.Index, client *mock.MockBeeClient) []byte {
+	ref, err := index.Delete(key)
+	if err != nil {
+		t.Fatal(err)
+	}
+	data, respCode, err := client.DownloadBlob(ref)
+	if err != nil {
+		t.Fatal(err)
+	}
+	if respCode != http.StatusOK {
+		t.Fatalf("invalid response code")
+	}
+	return data
+
+}
+
+func addLotOfDocs(t *testing.T, index *collection.Index, client *mock.MockBeeClient) map[string][]byte {
+	// Initialize the values
+	kvMap := make(map[string][]byte)
+	kvMap["key1"] = []byte("value1")
+	kvMap["key11"] = []byte("value11")
+	kvMap["aa"] = []byte("doc2")
+	kvMap["abc"] = []byte("doc4")
+	kvMap["abcd"] = []byte("doc5")
+	kvMap["aca"] = []byte("doc6")
+	kvMap["ac"] = []byte("doc7")
+	kvMap["acb"] = []byte("doc8")
+	kvMap["aaa"] = []byte("doc9")
+	kvMap["az"] = []byte("doc10")
+	kvMap["acdb"] = []byte("doc11")
+	kvMap["file1"] = []byte("doc12")
+	kvMap["file1.jpg"] = []byte("doc13")
+
+	// add the documents
+	for k, v := range kvMap {
+		addDoc(t, k, v, index, client)
+	}
+
+	kvMap["ab"] = []byte("doc3")
+	addDoc(t, "ab", kvMap["ab"], index, client)
+	kvMap["a"] = []byte("doc1")
+	addDoc(t, "a", kvMap["a"], index, client)
+	kvMap["key3"] = []byte("value3")
+	addDoc(t, "key3", kvMap["key3"], index, client)
+	kvMap["key2"] = []byte("value2")
+	addDoc(t, "key2", kvMap["key2"], index, client)
+
+	return kvMap
+}
+
+func addBatchDocs(t *testing.T, batch *collection.Batch, client *mock.MockBeeClient) map[string][]byte {
+	kvMap := make(map[string][]byte)
+	kvMap["key1"] = []byte("value1")
+	kvMap["key11"] = []byte("value11")
+	kvMap["aa"] = []byte("doc2")
+	kvMap["abc"] = []byte("doc4")
+	kvMap["abcd"] = []byte("doc5")
+	kvMap["aca"] = []byte("doc6")
+	kvMap["ac"] = []byte("doc7")
+	kvMap["acb"] = []byte("doc8")
+	kvMap["aaa"] = []byte("doc9")
+	kvMap["az"] = []byte("doc10")
+	kvMap["acdb"] = []byte("doc11")
+	kvMap["file1"] = []byte("doc12")
+	kvMap["file1.jpg"] = []byte("doc13")
+	kvMap["ab"] = []byte("doc3")
+	kvMap["a"] = []byte("doc1")
+	kvMap["key3"] = []byte("value3")
+	kvMap["key2"] = []byte("value2")
+	kvMap["ke77"] = []byte("batch doc1")
+	kvMap["ke79"] = []byte("batch doc2")
+	kvMap["a94"] = []byte("batch doc3")
+
+	// add the documents
+	for k, v := range kvMap {
+		ref, err := client.UploadBlob(v, false, false)
+		if err != nil {
+			t.Fatalf("could not add doc %s:%s, %v", k, ref, err)
+		}
+		err = batch.Put(k, ref)
+		if err != nil {
+			t.Fatal(err)
+		}
+	}
+	return kvMap
 }
